requirements.txt
# MCP Server for Code Documentation - Dependencies

# MCP Protocol
mcp>=1.0.0

# Code Analysis
astroid>=2.15.0
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-java>=0.20.0
tree-sitter-cpp>=0.20.0
tree-sitter-go>=0.20.0

# Natural Language Processing
nltk>=3.8.0
textstat>=0.7.0
spacy>=3.7.0

# Documentation Generation
jinja2>=3.1.0
markdown>=3.5.0
beautifulsoup4>=4.12.0

# File Processing
pathlib>=1.0.0
gitignore-parser>=0.1.0

# Configuration
pydantic>=2.5.0
pyyaml>=6.0.0
toml>=0.10.0

# Logging and Utilities
loguru>=0.7.0
click>=8.1.0
rich>=13.7.0

# Testing (development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0
